Add tests for formatting targets
