+++ .travis.yml
@@ -24,8 +24,8 @@ before_install:
   - cat ~/.bazelrc >> .bazelrc
 
 script:
-  - bazel build //...
-  - bazel test //...
+  - bazel build --strategy=Scalac=worker //...
+  - bazel test --strategy=Scalac=worker //...
   - bazel build src/scala/com/github/johnynek/bazel_deps/parseproject_deploy.jar
   - ./gen_maven_deps.sh `pwd` 3rdparty/workspace.bzl dependencies.yaml
   - git diff --exit-code
